+++ src/pv/loop.c
@@ -24,6 +24,32 @@
 #include <sys/ioctl.h>
 #include <sys/stat.h>
 
+#if HAVE_MATH_H
+#include <math.h>
+#endif
+
+
+#if HAVE_SQRTL
+#else
+/*
+ * Square root of a long double.  Adapted from iputils ping/ping_common.c.
+ */
+static long ldsqrt(long double value)
+{
+	long double previous = (long double) LLONG_MAX;
+	long double result = value;
+
+	if (result > 0) {
+		while (result < previous) {
+			previous = result;
+			result = (result + (value / result)) / 2;
+		}
+	}
+
+	return result;
+}
+#endif
+
 
 /*
  * Pipe data from a list of files to standard output, giving information
@@ -243,8 +269,11 @@ int pv_main_loop(pvstate_t state)
 			}
 		}
 
-		/* Just go round the loop again if there's no display. */
-		if (state->control.no_display)
+		/*
+		 * Just go round the loop again if there's no display and
+		 * we're not reporting statistics.
+		 */
+		if (state->control.no_display && !state->control.show_stats)
 			continue;
 
 		/*
@@ -331,7 +360,13 @@ int pv_main_loop(pvstate_t state)
 				state->control.height = new_height;
 		}
 
-		pv_display(state, final_update);
+		if (state->control.no_display) {
+			/* If there's no display, calculate rate for the statistics. */
+			pv_calculate_transfer_rate(state, final_update);
+		} else {
+			/* Produce the display. */
+			pv_display(state, final_update);
+		}
 	}
 
 	debug("%s: %s=%s, %s=%s", "loop ended", "eof_in", eof_in ? "true" : "false", "eof_out",
@@ -350,6 +385,52 @@ int pv_main_loop(pvstate_t state)
 
 	if (fd >= 0)
 		(void) close(fd);
+
+	/* Calculate and display the transfer statistics. */
+	if (state->control.show_stats && state->calc.measurements_taken > 0) {
+		char stats_buf[256];	 /* flawfinder: ignore */
+		long double rate_mean, rate_variance, rate_deviation;
+		int stats_size;
+
+		/* flawfinder: made safe by use of pv_snprintf() */
+
+		rate_mean = state->calc.rate_sum / ((long double) (state->calc.measurements_taken));
+		rate_variance =
+		    (state->calc.ratesquared_sum / ((long double) (state->calc.measurements_taken))) -
+		    (rate_mean * rate_mean);
+#if HAVE_SQRTL
+		rate_deviation = sqrtl(rate_variance);
+#else
+		rate_deviation = ldsqrt(rate_variance);
+#endif
+
+		debug("%s: %ld", "measurements taken", state->calc.measurements_taken);
+		debug("%s: %.3Lf", "rate_sum", state->calc.rate_sum);
+		debug("%s: %.3Lf", "ratesquared_sum", state->calc.ratesquared_sum);
+		debug("%s: %.3Lf", "rate_mean", rate_mean);
+		debug("%s: %.3Lf", "rate_variance", rate_variance);
+		debug("%s: %.3Lf", "rate_deviation", rate_deviation);
+
+		memset(stats_buf, 0, sizeof(stats_buf));
+		stats_size =
+		    pv_snprintf(stats_buf, sizeof(stats_buf), "%s = %.3Lf/%.3Lf/%.3Lf/%.3Lf %s\n",
+				_("rate min/avg/max/mdev"), state->calc.rate_min, rate_mean, state->calc.rate_max,
+				rate_deviation, state->control.bits ? _("b/s") : _("B/s"));
+
+		if (stats_size > 0 && stats_size < (int) (sizeof(stats_buf)))
+			pv_tty_write(state, stats_buf, (size_t) stats_size);
+	} else if (state->control.show_stats && state->calc.measurements_taken < 1) {
+		char msg_buf[256];	 /* flawfinder: ignore */
+		int msg_size;
+
+		/* flawfinder: made safe by use of pv_snprintf() */
+
+		memset(msg_buf, 0, sizeof(msg_buf));
+		msg_size = pv_snprintf(msg_buf, sizeof(msg_buf), "%s\n", _("rate not measured"));
+
+		if (msg_size > 0 && msg_size < (int) (sizeof(msg_buf)))
+			pv_tty_write(state, msg_buf, (size_t) msg_size);
+	}
 
 	return state->status.exit_status;
 }